Restructure these headers to use the same inner structs and avoid redefinition.
